This update includes several bugfixes
